move the `impl` parts to the corresponding view files
